% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_plotly_label}
\alias{add_plotly_label}
\title{Add Plotly Label}
\usage{
add_plotly_label(tbl, title, name, group)
}
\arguments{
\item{tbl}{A tibble}

\item{title}{A string}

\item{name}{Name of a column}

\item{group}{Name of a column}
}
\description{
Add Plotly Label
}
