% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_plotly_value_label}
\alias{add_plotly_value_label}
\title{Add Plotly Value Label}
\usage{
add_plotly_value_label(tbl, cols)
}
\arguments{
\item{tbl}{A tibble with column label}

\item{cols}{A vector of strings that are columns in the tibble}
}
\description{
Add Plotly Value Label
}
