% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo-sort.R
\name{bfs}
\alias{bfs}
\title{BFS of a graph.}
\usage{
bfs(adjlist, seeds)
}
\arguments{
\item{adjlist}{The adjacency list of the graph. See
\code{topologicalSort} for the expected format.}

\item{seeds}{Character vector, vertex ids to start the BFS from.}
}
\value{
Character vector with all vertex ids that are reachable
  from the \code{seeds}.
}
\description{
From the given vertices we collect all vertices that the reachable
from them. There are all the controls that we need to update.
}

