% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_manual_selection_functions.R
\name{build_cohort_object}
\alias{build_cohort_object}
\title{Build Cohort Object}
\usage{
build_cohort_object(
  dataset,
  samples,
  group_name,
  group_type,
  mutation_id = NA,
  bin_immune_feature = NA,
  bin_number = NA,
  filters = NA
)
}
\arguments{
\item{dataset}{A string}

\item{filter_obj}{A named list with element sample_ids and filters}

\item{group_choice}{A string}

\item{driver_mutation}{A string}

\item{immune_feature_bin_id}{An integer}

\item{immune_feature_bin_number}{An integer}
}
\description{
Build Cohort Object
}
