% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{build_cohort_tbl_by_feature_id}
\alias{build_cohort_tbl_by_feature_id}
\title{Build Cohort Tibble By Feature ID}
\usage{
build_cohort_tbl_by_feature_id(sample_ids, feature_id)
}
\arguments{
\item{sample_ids}{Integers in the id column of the samples table}

\item{feature_id}{An integer in the id column of the features_to_samples
table}
}
\description{
Build Cohort Tibble By Feature ID
}
