% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{build_cohort_tbl_by_group}
\alias{build_cohort_tbl_by_group}
\title{Build Cohort Table By Group}
\usage{
build_cohort_tbl_by_group(sample_ids, group)
}
\arguments{
\item{sample_ids}{Integers in the id column of the samples table}

\item{group}{A String that is the display column of the tags table}
}
\description{
Build Cohort Table By Group
}
