% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_feature_distributions_functions.R
\name{build_distplot_tbl}
\alias{build_distplot_tbl}
\title{Build Distribution Plot Tibble}
\usage{
build_distplot_tbl(sample_tbl, feature_id, scale_method)
}
\arguments{
\item{sample_tbl}{A tibble with columns "sample_id", "group"}

\item{feature_id}{An integer in the features table}

\item{scale_method}{a method for scaling the value column}
}
\description{
Build Distribution Plot Tibble
}
