% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{build_feature_bin_cohort_object}
\alias{build_feature_bin_cohort_object}
\title{Build Feature Bin Cohort Object}
\usage{
build_feature_bin_cohort_object(
  dataset,
  samples,
  feature_name,
  bin_number,
  feature_tbl
)
}
\arguments{
\item{bin_number}{An integer}

\item{sample_ids}{A vector of integers}

\item{feature_id}{An integer}
}
\description{
Build Feature Bin Cohort Object
}
