% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{build_feature_bin_sample_tbl}
\alias{build_feature_bin_sample_tbl}
\title{Build Feature Bin Sample Tibble}
\usage{
build_feature_bin_sample_tbl(dataset, samples, feature_name, n_bins)
}
\arguments{
\item{n_bins}{An integer}

\item{sample_ids}{A vector of integers}

\item{feature_id}{An integer}
}
\description{
Build Feature Bin Sample Tibble
}
