% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{build_feature_value_tbl}
\alias{build_feature_value_tbl}
\title{Build Feature Value Table}
\usage{
build_feature_value_tbl(feature_id)
}
\arguments{
\item{feature_id}{An integer in the feature_id column of the
features_to_samples table}
}
\description{
Build Feature Value Table
}
