% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{build_gene_expression_tbl_by_gene_ids}
\alias{build_gene_expression_tbl_by_gene_ids}
\title{Title}
\usage{
build_gene_expression_tbl_by_gene_ids(gene_ids)
}
\arguments{
\item{gene_ids}{Integers in the gene_id column of genes_to_samples}
}
\description{
Title
}
