% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_feature_distributions_functions.R
\name{build_histplot_tbl}
\alias{build_histplot_tbl}
\title{Build Histogram Tibble}
\usage{
build_histplot_tbl(distplot_tbl, group)
}
\arguments{
\item{distplot_tbl}{A tibble with columns "x", "y"}

\item{group}{A string whoose value is in the x column}
}
\description{
Build Histogram Tibble
}
