% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_feature_correlations_functions.R
\name{build_ifc_heatmap_matrix}
\alias{build_ifc_heatmap_matrix}
\title{Build Immune Feature Correlations Heatmap Matrix}
\usage{
build_ifc_heatmap_matrix(tbl, method)
}
\arguments{
\item{tbl}{A tibble with columns group, feature_name, order, feature_value,
and repsonse value}

\item{method}{A string, that is the correlaiton method}
}
\description{
Build Immune Feature Correlations Heatmap Matrix
}
