% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_feature_correlations_functions.R
\name{build_ifc_value_tbl}
\alias{build_ifc_value_tbl}
\title{Build Immune Feature Correlations Value Table}
\usage{
build_ifc_value_tbl(
  response_tbl,
  feature_tbl,
  sample_tbl,
  response_feature_name
)
}
\arguments{
\item{response_tbl}{A tibble with columns sample, name, value}

\item{feature_tbl}{A tibble with columns sample, feature_name,
feature_display, order, feature_value, tag}
}
\description{
Build Immune Feature Correlations Value Table
}
