% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunomodulators_functions.R
\name{build_im_distplot_tbl}
\alias{build_im_distplot_tbl}
\title{Build Immunomodulators Distributions Plot Tibble}
\usage{
build_im_distplot_tbl(cohort_object, gene, scale_method)
}
\arguments{
\item{scale_method}{A string, that is a scaling method}

\item{gene_id}{An integer in the gene_id column of the genes_types table}

\item{sample_tbl}{A tibble with columns sample, group}
}
\description{
Build Immunomodulators Distributions Plot Tibble
}
