% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunomodulators_functions.R
\name{build_im_dt_tbl}
\alias{build_im_dt_tbl}
\title{#' Build Immunomodulators Distributions Plot Tibble
#'
#' @param gene_id An integer in the gene_id column of the genes_types table
#' @param sample_tbl A tibble with columns sample, group
#' @param scale_method A string, that is a scaling method
#' @importFrom magrittr %>%
#' @importFrom dplyr select inner_join
#' @importFrom rlang .data
build_im_distplot_tbl <- function(gene, sample_tbl, scale_method){
    gene %>%
        as.integer() %>%
        build_gene_expression_tbl_by_gene_ids() %>%
        dplyr::inner_join(sample_tbl, by = "sample_id") %>%
        dplyr::select(.data$group, value = .data$rna_seq_expr) %>%
        scale_tbl_value_column(scale_method) %>%
        dplyr::select(x = .data$group, y = .data$value)
}
Build Immunomodulators Datatable Tibble}
\usage{
build_im_dt_tbl()
}
\arguments{
\item{tbl}{A tibble}
}
\description{
#' Build Immunomodulators Distributions Plot Tibble
#'
#' @param gene_id An integer in the gene_id column of the genes_types table
#' @param sample_tbl A tibble with columns sample, group
#' @param scale_method A string, that is a scaling method
#' @importFrom magrittr %>%
#' @importFrom dplyr select inner_join
#' @importFrom rlang .data
build_im_distplot_tbl <- function(gene, sample_tbl, scale_method){
    gene %>%
        as.integer() %>%
        build_gene_expression_tbl_by_gene_ids() %>%
        dplyr::inner_join(sample_tbl, by = "sample_id") %>%
        dplyr::select(.data$group, value = .data$rna_seq_expr) %>%
        scale_tbl_value_column(scale_method) %>%
        dplyr::select(x = .data$group, y = .data$value)
}
Build Immunomodulators Datatable Tibble
}
