% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_target_functions.R
\name{build_io_target_distplot_tbl}
\alias{build_io_target_distplot_tbl}
\title{Build IO Target Distplot Tibble}
\usage{
build_io_target_distplot_tbl(cohort_object, gene, scale_method)
}
\arguments{
\item{scale_method}{A string}

\item{gene_id}{An Integer in the gene_id column of the genes_to_samples table}

\item{sample_tbl}{A tibble with columns sample_id, and group}
}
\description{
Build IO Target Distplot Tibble
}
