% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_associations_multivariate_functions.R
\name{build_md_driver_violin_tbl}
\alias{build_md_driver_violin_tbl}
\title{Build Multivariate Driver Violin Tibble}
\usage{
build_md_driver_violin_tbl(tbl, .label)
}
\arguments{
\item{tbl}{A tibble with columns label, status and response}

\item{.label}{A string}
}
\description{
Build Multivariate Driver Violin Tibble
}
