% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{build_md_tag_covariate_tbl}
\alias{build_md_tag_covariate_tbl}
\title{Build Multivariate Driver Tag Covariate Tibble}
\usage{
build_md_tag_covariate_tbl(covariates)
}
\arguments{
\item{covariates}{A vector of strings that are in the name column of the tags
table}
}
\description{
Build Multivariate Driver Tag Covariate Tibble
}
