% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_functions.R
\name{build_numeric_filter_tbl}
\alias{build_numeric_filter_tbl}
\title{Build Numeric Filter Tibble}
\usage{
build_numeric_filter_tbl(feature_id)
}
\arguments{
\item{feature_id}{An integer in the feature_id column of the
features_to_samples table}
}
\description{
Build Numeric Filter Tibble
}
