% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/tumor_microenvironment_cell_proportions_functions.R
\name{build_ocp_barplot_tbl}
\alias{build_ocp_barplot_tbl}
\title{Build Overall Cell Proportions Barplot Tibble}
\usage{
build_ocp_barplot_tbl(value_tbl)
}
\arguments{
\item{value_tbl}{A tibble with columns feature_name, group, feature_value,
order}
}
\description{
Build Overall Cell Proportions Barplot Tibble
}
