% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_cell_proportions_functions.R
\name{build_ocp_scatterplot_tbl}
\alias{build_ocp_scatterplot_tbl}
\title{Build Overall Cell Proportions Scatterplot Tibble}
\usage{
build_ocp_scatterplot_tbl(value_tbl, group_value)
}
\arguments{
\item{value_tbl}{A tibble with columns sample_id, group, feature_name,
feature_value}

\item{group_value}{A string, that is in the group column of the value_tbl}
}
\description{
Build Overall Cell Proportions Scatterplot Tibble
}
