% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_cell_proportions_functions.R
\name{build_ocp_value_tbl}
\alias{build_ocp_value_tbl}
\title{Build Overall Cell Proportions Value Tibble}
\usage{
build_ocp_value_tbl(sample_tbl)
}
\arguments{
\item{sample_tbl}{A tibble with columns sample_id and group}
}
\description{
Build Overall Cell Proportions Value Tibble
}
