% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_outcomes_survival_functions.R
\name{build_survival_values_tbl}
\alias{build_survival_values_tbl}
\title{Build Survival Values Tibble}
\usage{
build_survival_values_tbl(sample_tbl, time_id, status_id)
}
\arguments{
\item{sample_tbl}{A Tibble with columns sample_id and group}

\item{time_id}{An integer in the id column of the samples table}

\item{status_id}{An integer in the id column of the samples table}
}
\description{
Build Survival Values Tibble
}
