% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/til_map_datatable_functions.R
\name{build_tm_dt_tbl}
\alias{build_tm_dt_tbl}
\title{Build Tilmap Datatbale Tibble}
\usage{
build_tm_dt_tbl(sample_tbl)
}
\arguments{
\item{sample_tbl}{A tibble with columns sample_id, sample_name, group,
slide_barcode}
}
\description{
Build Tilmap Datatbale Tibble
}
