% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/til_maps_functions.R
\name{build_tm_sample_tbl}
\alias{build_tm_sample_tbl}
\title{Build Tilmap Sample Tibble}
\usage{
build_tm_sample_tbl(sample_tbl)
}
\arguments{
\item{sample_tbl}{A tibble with columns sample_id and group}
}
\description{
Build Tilmap Sample Tibble
}
