% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_driver_functions.R
\name{calculate_ratio_effect_size}
\alias{calculate_ratio_effect_size}
\title{Calculate Ratio Effect Size}
\usage{
calculate_ratio_effect_size(v1, v2)
}
\arguments{
\item{v1}{A numeric vector}

\item{v2}{A numeric vector}
}
\description{
Calculate Ratio Effect Size
}
