% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_associations_multivariate_functions.R
\name{combine_md_tbls}
\alias{combine_md_tbls}
\title{Combine Multivariate Driver Tibbles}
\usage{
combine_md_tbls(resp_tbl, status_tbl, sample_tbl, cov_tbl, mode)
}
\arguments{
\item{resp_tbl}{A tibble}

\item{status_tbl}{A tibble}

\item{sample_tbl}{A tibble}

\item{cov_tbl}{A tibble or NULL}

\item{mode}{A string, either "By group" or "Across groups"}
}
\description{
Combine Multivariate Driver Tibbles
}
