% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Create a database connection}
\usage{
connect_to_db(
  dbname = .GlobalEnv$DB_NAME,
  host = .GlobalEnv$DB_HOST,
  port = .GlobalEnv$DB_PORT,
  user = .GlobalEnv$DB_USER,
  password = .GlobalEnv$DB_PW
)
}
\arguments{
\item{dbname}{String}

\item{host}{String}

\item{port}{String}

\item{user}{String}

\item{password}{String}
}
\description{
Create a database connection
}
