% Generated by roxygen2 (5.0.0.9000): do not edit by hand
% Please edit documentation in R/controlTypes.R
\docType{data}
\name{control_classes}
\alias{control_classes}
\title{Front-end control types and matching back-end classes}
\format{\preformatted{List of 18
 $ rplot           :Class 'R6ClassGenerator' <environment: 0x7fbef32fbd50> 
  ..- attr(*, "name")= chr "RPlotControl_generator"
 $ interactiveplot :Class 'R6ClassGenerator' <environment: 0x7fbef32de6e8> 
  ..- attr(*, "name")= chr "InteractivePlotControl_generator"
 $ form            :Class 'R6ClassGenerator' <environment: 0x7fbef30bdf80> 
  ..- attr(*, "name")= chr "FormControl_generator"
 $ textfield       :Class 'R6ClassGenerator' <environment: 0x7fbef32bf550> 
  ..- attr(*, "name")= chr "TextFieldControl_generator"
 $ datepicker      :Class 'R6ClassGenerator' <environment: 0x7fbef3290cd0> 
  ..- attr(*, "name")= chr "DatePickerControl_generator"
 $ dropdown        :Class 'R6ClassGenerator' <environment: 0x7fbef3261548> 
  ..- attr(*, "name")= chr "DropdownControl_generator"
 $ multiselect     :Class 'R6ClassGenerator' <environment: 0x7fbef3160578> 
  ..- attr(*, "name")= chr "MultiSelectControl_generator"
 $ checkboxlist    :Class 'R6ClassGenerator' <environment: 0x7fbef730c6a0> 
  ..- attr(*, "name")= chr "CheckboxListControl_generator"
 $ radiobuttongroup:Class 'R6ClassGenerator' <environment: 0x7fbef7724f10> 
  ..- attr(*, "name")= chr "RadioButtonGroupControl_generator"
 $ slider          :Class 'R6ClassGenerator' <environment: 0x7fbef76c6308> 
  ..- attr(*, "name")= chr "Slidercontrol_generator"
 $ separator       :Class 'R6ClassGenerator' <environment: 0x7fbef7188d88> 
  ..- attr(*, "name")= chr "SeparatorControl_generator"
 $ heading         :Class 'R6ClassGenerator' <environment: 0x7fbef70974d8> 
  ..- attr(*, "name")= chr "HeadingControl_generator"
 $ submitbutton    :Class 'R6ClassGenerator' <environment: 0x7fbef752ab08> 
  ..- attr(*, "name")= chr "SubmitButtonControl_generator"
 $ iframe          :Class 'R6ClassGenerator' <environment: 0x7fbef7517110> 
  ..- attr(*, "name")= chr "IFrameControl_generator"
 $ image           :Class 'R6ClassGenerator' <environment: 0x7fbef7508f18> 
  ..- attr(*, "name")= chr "ImageControl_generator"
 $ pagemenu        :Class 'R6ClassGenerator' <environment: 0x7fbef74f4520> 
  ..- attr(*, "name")= chr "PageMenuControl_generator"
 $ breadcrumb      :Class 'R6ClassGenerator' <environment: 0x7fbef74e32d0> 
  ..- attr(*, "name")= chr "BreadCrumbControl_generator"
 $ text            :Class 'R6ClassGenerator' <environment: 0x7fbef6849498> 
  ..- attr(*, "name")= chr "TextControl_generator"
}}
\usage{
control_classes
}
\description{
Front-end control types and matching back-end classes
}
\keyword{datasets}

