% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_build_get_gene_expression_tbl_by_gene_ids_query}
\alias{create_build_get_gene_expression_tbl_by_gene_ids_query}
\title{Create Build Get Gene Expression Table By Gene IDs Query}
\usage{
create_build_get_gene_expression_tbl_by_gene_ids_query(gene_ids)
}
\arguments{
\item{gene_ids}{Integers in the gene_id column of the genes_to_samples table}
}
\description{
Create Build Get Gene Expression Table By Gene IDs Query
}
