% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{create_categorical_covariate_string}
\alias{create_categorical_covariate_string}
\title{Create Categorical Covariate String}
\usage{
create_categorical_covariate_string(covs, translate_func)
}
\arguments{
\item{covs}{A vector of strings or NULL}

\item{translate_func}{A function}
}
\description{
Create Categorical Covariate String
}
