% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_dm_cohort_object}
\alias{create_dm_cohort_object}
\title{Create Driver Mutation Cohort Object}
\usage{
create_dm_cohort_object(sample_ids, mutation)
}
\arguments{
\item{sample_ids}{A vector of integers}

\item{mutation}{A string}
}
\description{
Create Driver Mutation Cohort Object
}
