% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_feature_bin_cohort_object}
\alias{create_feature_bin_cohort_object}
\title{Create Feature Bin Cohort Object}
\usage{
create_feature_bin_cohort_object(sample_ids, feature_id, bin_number)
}
\arguments{
\item{sample_ids}{A vector of integers}

\item{feature_id}{An integer}

\item{bin_number}{An integer}
}
\description{
Create Feature Bin Cohort Object
}
