% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_feature_bin_group_tbl}
\alias{create_feature_bin_group_tbl}
\title{Create Feature Bin Group Tibble}
\usage{
create_feature_bin_group_tbl(sample_tbl, feature_name)
}
\arguments{
\item{sample_tbl}{A tibble with name group}

\item{feature_name}{A string}
}
\description{
Create Feature Bin Group Tibble
}
