% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_feature_named_list}
\alias{create_feature_named_list}
\title{Create Feature Named List}
\usage{
create_feature_named_list(class_ids = "all")
}
\arguments{
\item{class_ids}{Integers in the id column of the classes table}
}
\description{
Create Feature Named List
}
