% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_functions.R
\name{create_feature_tbl}
\alias{create_feature_tbl}
\title{Create Feature Tibble}
\usage{
create_feature_tbl(class_choice_id)
}
\arguments{
\item{class_choice_id}{An integer in class_id column of the features table,
or -1}
}
\description{
Create Feature Tibble
}
