% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_feature_value_query}
\alias{create_feature_value_query}
\title{Create Feature Value Query}
\usage{
create_feature_value_query(subquery)
}
\arguments{
\item{subquery}{One of :
- A subquery that results in a list of feature ids
- A subquery that results in a one column table of feature ids
- A single feature_id}
}
\description{
Create Feature Value Query
}
