% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_feature_value_query_from_class_ids}
\alias{create_feature_value_query_from_class_ids}
\title{Create Feature Value Query from class ids}
\usage{
create_feature_value_query_from_class_ids(class_ids)
}
\arguments{
\item{class_ids}{class ids in the features to samples table}
}
\description{
Create Feature Value Query from class ids
}
