% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_feature_value_query_from_subquery}
\alias{create_feature_value_query_from_subquery}
\title{Create Feature Value Query from subquery}
\usage{
create_feature_value_query_from_subquery(subquery)
}
\arguments{
\item{subquery}{subquery that results in a list of feature ids}
}
\description{
Create Feature Value Query from subquery
}
