% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_get_class_id_from_name_query}
\alias{create_get_class_id_from_name_query}
\title{Create Get Class Id from Name Query}
\usage{
create_get_class_id_from_name_query(name)
}
\arguments{
\item{name}{A feature's name}
}
\description{
Create Get Class Id from Name Query
}
