% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_get_feature_id_from_display_query}
\alias{create_get_feature_id_from_display_query}
\title{Create Get Feature Id from Display Query}
\usage{
create_get_feature_id_from_display_query(display)
}
\arguments{
\item{display}{A feature display name}
}
\description{
Create Get Feature Id from Display Query
}
