% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_get_genes_by_type_query}
\alias{create_get_genes_by_type_query}
\title{Create Get Genes by Type Query}
\usage{
create_get_genes_by_type_query(gene_type)
}
\arguments{
\item{gene_type}{The name of the gene type}
}
\description{
Create Get Genes by Type Query
}
