% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_get_sample_ids_from_parent_tag_display_query}
\alias{create_get_sample_ids_from_parent_tag_display_query}
\title{Create Get Sample IDs From Parent Tag Display Query}
\usage{
create_get_sample_ids_from_parent_tag_display_query(display)
}
\arguments{
\item{display}{A string in the display column of the tags table}
}
\description{
Create Get Sample IDs From Parent Tag Display Query
}
