% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline_functions.R
\name{create_heritability_df}
\alias{create_heritability_df}
\title{Prepare heritability tibble for plotting}
\usage{
create_heritability_df(
  heritablity_data,
  parameter = "cluster",
  group = "European",
  pval_thres = 0.05,
  ancestry_labels
)
}
\arguments{
\item{heritablity_data}{A tibble}

\item{parameter}{Parameter to be used for selection of results (eg, ancestry cluster, immune feature)}

\item{group}{Specific group, in the selected parameter, to be displayed (eg. European - ancestry cluster, NK cells - immune feature)}

\item{pval_thres}{Maximun p-value to be included}

\item{ancestry_labels}{A list}
}
\description{
Prepare heritability tibble for plotting
}
