% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunomodulators_functions.R
\name{create_im_gene_list}
\alias{create_im_gene_list}
\title{Create Immunomodulators Gene List}
\usage{
create_im_gene_list(tbl, group)
}
\arguments{
\item{tbl}{A tibble with columns hgnc, id, and the variable group}

\item{group}{A column in the tibble}
}
\description{
Create Immunomodulators Gene List
}
