% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_target_functions.R
\name{create_io_target_gene_list}
\alias{create_io_target_gene_list}
\title{Create IO Target Gene List}
\usage{
create_io_target_gene_list(tbl, group)
}
\arguments{
\item{tbl}{A Tibble with columns hgnc, id, and the value of the group var}

\item{group}{A Column in the tibble}
}
\description{
Create IO Target Gene List
}
