% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_manual_selection_functions.R
\name{create_mutation_cohort_group_tbl}
\alias{create_mutation_cohort_group_tbl}
\title{Create Driver Mutation Cohort Group Tibble}
\usage{
create_mutation_cohort_group_tbl(sample_tbl, mutation)
}
\arguments{
\item{sample_tbl}{A tibble}

\item{gene_name}{A string}
}
\description{
Create Driver Mutation Cohort Group Tibble
}
