% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_queries.R
\name{create_parent_group_query_from_display}
\alias{create_parent_group_query_from_display}
\title{Create Parent Group Query From Display}
\usage{
create_parent_group_query_from_display(display)
}
\arguments{
\item{display}{The display name of the parent group}
}
\description{
Create Parent Group Query From Display
}
