% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_plotly_label}
\alias{create_plotly_label}
\title{Create Plotly Label}
\usage{
create_plotly_label(tbl, name, group, cols, title = "ParticipantBarcode")
}
\arguments{
\item{tbl}{A tibble}

\item{name}{A column}

\item{group}{A column}

\item{cols}{A vector of strings, whioch are columns in the tibble}

\item{title}{A string}
}
\description{
Create Plotly Label
}
