% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{create_tag_cohort_object}
\alias{create_tag_cohort_object}
\title{Create Tag Cohort Object}
\usage{
create_tag_cohort_object(samples, dataset, tag)
}
\arguments{
\item{sample_ids}{A vector of integers}

\item{group_choice}{A string}
}
\description{
Create Tag Cohort Object
}
