% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/til_map_datatable_functions.R
\name{create_tm_slide_link}
\alias{create_tm_slide_link}
\title{Create Tilmap Slide Link}
\usage{
create_tm_slide_link(slide_barcodes)
}
\arguments{
\item{slide_barcodes}{A vector of strings}
}
\description{
Create Tilmap Slide Link
}
