% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-server.R
\name{cria_da_incos}
\alias{cria_da_incos}
\title{Estrutura as bases de dados}
\usage{
cria_da_incos(lista_bases, session)
}
\arguments{
\item{lista_bases}{lista contendo assuntos e basics}

\item{session}{session do shiny}
}
\description{
Lê os assuntos e dados basicos (saida de parse_files) e retorna o da_incos
}
