% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-server.R
\name{cria_da_totais}
\alias{cria_da_totais}
\title{Estrutura as bases de dados (da_totais)}
\usage{
cria_da_totais(lista_bases, da_incos)
}
\arguments{
\item{lista_bases}{lista contendo assuntos e basics}

\item{da_incos}{base de inconsistencias}
}
\description{
Lê o da_incos e o basics e cria o da_totais
}
