% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_incos}
\alias{da_incos}
\title{da_incos}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100000 rows and 25 columns.
}
\usage{
da_incos
}
\description{
da_incos
}
\keyword{datasets}
