% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_sample}
\alias{da_sample}
\title{da_sample}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 24 columns.
}
\usage{
da_sample
}
\description{
da_sample
}
\keyword{datasets}
