% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_totais}
\alias{da_totais}
\title{da_totais}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 89 rows and 7 columns.
}
\usage{
da_totais
}
\description{
da_totais
}
\keyword{datasets}
