% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fillTemplate.R
\name{fillTemplate}
\alias{fillTemplate}
\title{Fill in a function template}
\usage{
fillTemplate(templateName = NULL, funParams)
}
\arguments{
\item{templateName}{The file name (without directory) of the template. This
should be stored in \code{inst/Rtemplates}.}

\item{funParams}{A named list. The item name is the variable name in the
template. The item value is what it should be replaced with.}
}
\value{
Character vector with the filled in template, each line with its own
entry in the vector.
}
\description{
The template is an R function with variables between tags with a
<%=VARIABLENAME%> format. On calling this function all the variables are
replaced with the supplied values and the function returned as a vector of
lines.
}
\examples{
funText <- fillTemplate("rplotTemplate.R",
                        list(control.id="rcapid123",
                             control.controlProperties.1="hist(rnorm(1000))"))
cat(paste(funText, collapse="\\n"))
}

