% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_functions.R
\name{filter_feature_tbl}
\alias{filter_feature_tbl}
\title{Filter Feature Tibble}
\usage{
filter_feature_tbl(feature_tbl, selected_row)
}
\arguments{
\item{feature_tbl}{A tibble with columns class_id}

\item{selected_row}{An integer that is a row number in the tibble}
}
\description{
Filter Feature Tibble
}
