% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_associations_multivariate_functions.R
\name{filter_md_labels}
\alias{filter_md_labels}
\title{Filter Multivariate Driver Labels}
\usage{
filter_md_labels(tbl, min_mutants, min_wildtype)
}
\arguments{
\item{tbl}{A tibble}

\item{min_mutants}{A positive integer}

\item{min_wildtype}{A positive integer}
}
\description{
Filter Multivariate Driver Labels
}
