% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info_functions.R
\name{format_feature_tbl}
\alias{format_feature_tbl}
\title{Format Feature Tibble}
\usage{
format_feature_tbl(feature_tbl)
}
\arguments{
\item{feature_tbl}{A tibble with columns feature, unit and class}
}
\description{
Format Feature Tibble
}
