% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{func_montante_previsto}
\alias{func_montante_previsto}
\title{Calculo de entradas e saidas}
\usage{
func_montante_previsto(fluxo, futuro, tipo = "entrada")
}
\arguments{
\item{fluxo}{tabela fluxo de caixa}

\item{futuro}{tabela futuro}

\item{tipo}{tipo de visualizacao}
}
\description{
Calculo de entradas e saidas
}
