% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{func_saldo_fim_mes}
\alias{func_saldo_fim_mes}
\title{Calculo do saldo no fim do mes}
\usage{
func_saldo_fim_mes(fluxo, futuro)
}
\arguments{
\item{fluxo}{tabela fluxo de caixa}

\item{futuro}{tabela futuro}
}
\description{
Calculo do saldo no fim do mes
}
