% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_class_id_from_name}
\alias{get_class_id_from_name}
\title{Get Class ID from Class Name}
\usage{
get_class_id_from_name(name)
}
\arguments{
\item{name}{An string in the name column of the classes table}
}
\description{
Get Class ID from Class Name
}
