% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_feature_id_from_display}
\alias{get_feature_id_from_display}
\title{Get Feature ID From Display Name}
\usage{
get_feature_id_from_display(display)
}
\arguments{
\item{display}{A string in the display column of the features table}
}
\description{
Get Feature ID From Display Name
}
