% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_feature_values_from_ids}
\alias{get_feature_values_from_ids}
\title{Get Feature Values Tibble Name From IDs}
\usage{
get_feature_values_from_ids(ids)
}
\arguments{
\item{ids}{An Integers in the id column of the features_to_samples table}
}
\description{
Get Feature Values Tibble Name From IDs
}
