% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{get_filtered_group_sample_ids}
\alias{get_filtered_group_sample_ids}
\title{Get Filtered Group Sample IDs}
\usage{
get_filtered_group_sample_ids(filter_obj, sample_ids)
}
\arguments{
\item{filter_obj}{A list of named lists with names ids and name}

\item{sample_ids}{A vector on integers that are in the sample_id column of
samples_to_tags table}
}
\description{
Get Filtered Group Sample IDs
}
