% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_group_selection_functions.R
\name{get_gene_from_mutation}
\alias{get_gene_from_mutation}
\title{Get Gene From Mutation}
\usage{
get_gene_from_mutation(mutation)
}
\arguments{
\item{mutation}{A string}
}
\description{
Get Gene From Mutation
}
