% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_id_from_mutation_code}
\alias{get_id_from_mutation_code}
\title{Get ID From Mutation Code}
\usage{
get_id_from_mutation_code(code)
}
\arguments{
\item{code}{A string in the code column of the mutation_codes table}
}
\description{
Get ID From Mutation Code
}
