% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{get_items_from_numeric_covariate_output}
\alias{get_items_from_numeric_covariate_output}
\title{Get Items From Numeric Covariate Output}
\usage{
get_items_from_numeric_covariate_output(output, name)
}
\arguments{
\item{output}{A named list of named lists}

\item{name}{A name in all of the named lists}
}
\description{
Get Items From Numeric Covariate Output
}
