% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_functions.R
\name{get_names_from_categorical_covariate_output}
\alias{get_names_from_categorical_covariate_output}
\title{Get Names From Categorical Covariate Output}
\usage{
get_names_from_categorical_covariate_output(output)
}
\arguments{
\item{output}{A named list of named lists}
}
\description{
Get Names From Categorical Covariate Output
}
