% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_outcomes_functions.R
\name{get_status_id_from_time_id}
\alias{get_status_id_from_time_id}
\title{Get Survival Status ID from Time ID}
\usage{
get_status_id_from_time_id(time_id)
}
\arguments{
\item{time_id}{An integer of a time status feature}
}
\description{
Get Survival Status ID from Time ID
}
