% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_tag_display_from_id}
\alias{get_tag_display_from_id}
\title{Get Tag Display Name from ID}
\usage{
get_tag_display_from_id(id)
}
\arguments{
\item{id}{An integer in the id column of the tags table}
}
\description{
Get Tag Display Name from ID
}
