% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_queries.R
\name{get_tag_display_from_name}
\alias{get_tag_display_from_name}
\title{Get Tag Display Name from Name}
\usage{
get_tag_display_from_name(name)
}
\arguments{
\item{name}{A string in the name column of the tags table}
}
\description{
Get Tag Display Name from Name
}
