% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_values_from_col}
\alias{get_unique_values_from_col}
\title{Get Unique Values from Column}
\usage{
get_unique_values_from_col(tbl, col)
}
\arguments{
\item{tbl}{A tibble}

\item{col}{A column in the tibble}
}
\description{
Get Unique Values from Column
}
