% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{Title}
\usage{
get_url(
  q,
  url = Sys.getenv("urlapi", "127.0.0.1"),
  port = Sys.getenv("portapi", "8080")
)
}
\arguments{
\item{q}{param pour l'url}

\item{url}{url de l'api}

\item{port}{port de lapi}
}
\description{
Title
}
