% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_filter_selection_functions.R
\name{get_valid_numeric_filters}
\alias{get_valid_numeric_filters}
\title{Get Valid Numeric Filters}
\usage{
get_valid_numeric_filters(filter_obj)
}
\arguments{
\item{filter_obj}{A list of named lists with names id max and min}
}
\description{
Get Valid Numeric Filters
}
