% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_values_from_eventdata}
\alias{get_values_from_eventdata}
\title{Get Values from Eventdata Dataframe}
\usage{
get_values_from_eventdata(eventdata, col = "x")
}
\arguments{
\item{eventdata}{Eventdata from "plotly_click" plotly::event_data}

\item{col}{The column to get the values from}
}
\description{
Get Values from Eventdata Dataframe
}
