% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getControls.R
\name{getControls}
\alias{getControls}
\title{Return a list of all controls}
\usage{
getControls(rcapConfig)
}
\arguments{
\item{rcapConfig}{a list generated by jsonlite::fromJSON(json, simplifyVector = TRUE)}

\item{rcapPage}{a list containing the controls on a page}
}
\value{
List of controls (each a list)


list of the controls (each a list) on the page and the functions of child pages in
a tree structure.

}
\description{
Process and return each control that creates an OCAP R function 
on a page. This function will call itself recursively for all child pages in
a depth first search.

}
\examples{
Parse an RCAP page for functions

}

