% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_gitminer.r
\docType{class}
\name{gitminer-class}
\alias{GitMiner}
\alias{gitminer-class}
\title{gitminerで収集したメトリクスPlotを行うクラス
コンストラクタ}
\arguments{
\item{datasource}{データソースが保持されているURL(CSV形式)
箱ひげ図をプロットする関数}

\item{query}{フィルタリング文字列}

\item{f}{箱ひげ図に表示するためのデータを生成するためにメトリクスデータに対して実行される関数
全メトリクスデータから、filter各プロジェクトの最新データのみをフィルタリングし、fieldで指定されたデータ項目を総行数で割った値（密度）を返す関数}

\item{group}{フィルタリング文字列}

\item{field}{密度を求めるメトリクス項目
全メトリクスデータから、各プロジェクトの最新データのみをフィルタリングして返す関数}

\item{group}{フィルタリング文字列}

\item{group}{フィルタリング文字列}

\item{data}{密度を求める対象のメトリクスデータ}

\item{field}{密度を求める対象のメトリクスデータ項目}
}
\value{
各プロジェクトの最新メトリクス情報
全メトリクスデータから引数groupで指定されたプロジェクトのデータのみを返す関数

引数のgroupでフィルタリングされたメトリクスデータ
引数で与えられたfieldの値を総行数で割った密度を求める関数

引数dataのfieldで指定された項目を総行数で割った値
メトリクスデータの項目名を返す関数

メトリクスデータの項目名リスト
}
\description{
gitminerで収集したメトリクスPlotを行うクラス
コンストラクタ
}

