% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-asset.R
\name{graf_entradas_saidas}
\alias{graf_entradas_saidas}
\title{Gráfico de entradas e saídas}
\usage{
graf_entradas_saidas(fluxo, futuro, visao = "dia")
}
\arguments{
\item{fluxo}{tabela fluxo de caixa}

\item{futuro}{tabela futuro}

\item{visao}{tipo de visualizacao}
}
\description{
Gráfico de entradas e saídas
}
