% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-asset.R
\name{graf_fluxo_caixa}
\alias{graf_fluxo_caixa}
\title{Gráfico do fluxo de caixa}
\usage{
graf_fluxo_caixa(fluxo, futuro, visao = "dia")
}
\arguments{
\item{fluxo}{tabela fluxo de caixa}

\item{futuro}{tabela futuro}

\item{visao}{tipo de visualizacao}
}
\description{
Gráfico do fluxo de caixa
}
