% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_04_imputations.R
\name{imp_descr}
\alias{imp_descr}
\title{Title Compute descriptive statistics for imputed data}
\usage{
imp_descr(imp)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}
}
\value{
A datatable with descriptives
}
\description{
Title Compute descriptive statistics for imputed data
}
