% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_incos_fun.R
\name{inc_numero_justica_tribunal_fun}
\alias{inc_numero_justica_tribunal_fun}
\alias{inc_classe_fun}
\alias{incos}
\alias{inc_digito_fun}
\alias{inc_data_ajuizamento_fun}
\alias{inc_sistema_fun}
\alias{inc_proc_el_fun}
\alias{inc_valor_fun}
\alias{inc_orgao_fun}
\alias{inc_municipio_fun}
\alias{inc_assuntos_fun}
\alias{inc_classe_assunto_fun}
\alias{inc_mov_responsavel_fun}
\alias{inc_mov_demorada_fun}
\alias{inc_mov_cod_pai_faltante_fun}
\alias{inc_mov_processo_longo_fun}
\alias{inc_mov_id_segue_ordem_cronologica_fun}
\alias{inc_mov_relevante_ordenada_fun}
\alias{inc_mov_relevante_faltante_fun}
\title{inc_numero_justica_tribunal_fun}
\usage{
inc_numero_justica_tribunal_fun(da)

inc_classe_fun(da)

inc_digito_fun(da)

inc_data_ajuizamento_fun(da)

inc_sistema_fun(da)

inc_proc_el_fun(da)

inc_valor_fun(da)

inc_orgao_fun(da)

inc_municipio_fun(da)

inc_assuntos_fun(da_assunto, sgt_assunto)

inc_classe_assunto_fun(da_assunto, da_basic, sgt_assunto)

inc_mov_responsavel_fun(mov)

inc_mov_demorada_fun(mov)

inc_mov_cod_pai_faltante_fun(mov)

inc_mov_processo_longo_fun(mov)

inc_mov_id_segue_ordem_cronologica_fun(mov)

inc_mov_relevante_ordenada_fun(mov)

inc_mov_relevante_faltante_fun(mov)
}
\arguments{
\item{da}{base de dados}

\item{da_assunto}{base de assuntos}

\item{sgt_assunto}{assuntos oficiais CNJ}

\item{da_basic}{base de dados basicos}

\item{mov}{movimentacao}
}
\description{
funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"

funções inc_ geram as inconsistências da base "da"
}
