% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_tbl_value_column}
\alias{log_tbl_value_column}
\title{Log Tibble Value Column}
\usage{
log_tbl_value_column(tbl, base = 10, add_amt = 0)
}
\arguments{
\item{tbl}{A Tibble with column "value"}

\item{base}{An integer, used as the base in log}

\item{add_amt}{A numeric, added to the vlaue column before logging}
}
\description{
Log Tibble Value Column
}
