% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{mapa}
\alias{mapa}
\title{mapa}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5572 rows and 7 columns.
}
\usage{
mapa
}
\description{
mapa
}
\keyword{datasets}
