% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_05_save.R
\name{mids2spss}
\alias{mids2spss}
\title{Title Save imputations in SPSS readable format}
\usage{
mids2spss(imp, file)
}
\arguments{
\item{imp}{A multiply imputed data set (mids) object}

\item{file}{Temporary file location needed for shiny::downloadHandler()}
}
\value{
A SPPS readable multiply imputed data set
}
\description{
Title Save imputations in SPSS readable format
}
