% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_authentification.R
\name{mod_authentification_server}
\alias{mod_authentification_server}
\title{authentification Server Functions}
\usage{
mod_authentification_server(id)
}
\description{
authentification Server Functions
}
