% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_geral.R
\name{mod_geral_server}
\alias{mod_geral_server}
\title{geral Server Functions}
\usage{
mod_geral_server(id, app_data)
}
\arguments{
\item{id, app_data}{Internal parameters for {shiny}.}
}
\description{
geral Server Functions
}
