% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_selectInput_GET_compartment_replicates.R
\name{mod_selectInput_GET_compartment_replicates_server}
\alias{mod_selectInput_GET_compartment_replicates_server}
\title{selectInput_GET_compartment_replicates Server Functions}
\usage{
mod_selectInput_GET_compartment_replicates_server(
  id,
  authentification_module,
  api_function_options = list(),
  widget_options = list()
)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{authentification_module}{Authentification module from this package}

\item{api_function_options}{List of options to pass to GET /compartment_replicates as query parameters}
}
\value{

}
\description{
selectInput_GET_compartment_replicates Server Functions
}
