% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_selectInput_GET_projects.R
\name{mod_selectInput_GET_projects_ui}
\alias{mod_selectInput_GET_projects_ui}
\title{selectInput_GET_projects UI Function}
\usage{
mod_selectInput_GET_projects_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around the deepomics' API GET projects call.
}
