% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summary.R
\name{montar_tabela_caixa}
\alias{montar_tabela_caixa}
\title{monta reactable da tabela de extrato}
\usage{
montar_tabela_caixa(tab)
}
\arguments{
\item{tab}{tabela}
}
\description{
monta reactable da tabela de extrato
}
