% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-asset.R
\name{my_infobox}
\alias{my_infobox}
\title{Infobox customizado
Todos os parametros da funcao shiny::infobox, mas com 2 a mais, dos quais}
\usage{
my_infobox(
  title,
  value = NULL,
  subtitle = NULL,
  icon = shiny::icon("bar-chart"),
  bg_color = "#224563",
  color_icon = "white",
  width = 4
)
}
\arguments{
\item{title}{titulo}

\item{value}{valor da caixinha}

\item{subtitle}{subtitulo}

\item{icon}{icone}

\item{bg_color}{cor do background do icone}

\item{color_icon}{cor do icone}

\item{width}{width}
}
\description{
Infobox customizado
Todos os parametros da funcao shiny::infobox, mas com 2 a mais, dos quais
}
