% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopcat-package.R
\docType{package}
\name{omopcat-package}
\alias{omopcat}
\alias{omopcat-package}
\title{omopcat: The OMOP Data Catalogue Dashboard}
\description{
A Shiny app to display an OMOP data catalogue in an interactive dashboard. Enables users to interactively explore available OMOP concepts by displaying useful summary statistics and subsequently export a selection of concepts of interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SAFEHR-data/omopcat}
  \item Report bugs at \url{https://github.com/SAFEHR-data/omopcat/issues}
}

}
\author{
\strong{Maintainer}: Milan Malfait \email{m.malfait@ucl.ac.uk}

Authors:
\itemize{
  \item Stefan Piatek
  \item Baptiste Briot Ribeyre
  \item Andy South
}

Other contributors:
\itemize{
  \item SAFEHR [copyright holder, funder]
}

}
\keyword{internal}
