% Generated by roxygen2 (5.0.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{or}
\alias{\%||\%}
\alias{or}
\title{Evaluate rhs if lhs is NULL}
\usage{
l \%||\% r
}
\arguments{
\item{l}{Left hand side, evaluate this first.}

\item{r}{Right had side, only evaluate it if \code{l} is \code{NULL}.}
}
\value{
\code{l} if it is not \code{NULL}, otherwise \code{r}.
}
\description{
Evaluate rhs if lhs is NULL
}

