% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{orgaos_base}
\alias{orgaos_base}
\title{orgaos_base}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39096 rows and 1 columns.
}
\usage{
orgaos_base
}
\description{
orgaos_base
}
\keyword{datasets}
