% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-server.R
\name{parse_file}
\alias{parse_file}
\title{Import files}
\usage{
parse_file(infile, names)
}
\arguments{
\item{infile}{Arquivo .json ou .xml a ser estruturado}

\item{names}{nomes dos arquivos}
}
\description{
Lê um arquivo .json ou .xml e retorna uma base de dados pronta para o uso
}
