% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.R
\name{parseConfig}
\alias{parseConfig}
\title{Parse Configuration List}
\usage{
parseConfig(rcapConfig)
}
\arguments{
\item{rcapConfig}{a list generated by jsonlite::fromJSON(json, simplifyVector = TRUE)}
}
\value{
List of R function names in same hierachy as page structure
}
\description{
Parse Configuration List
}
\examples{
jsonFile <- system.file(package = "rcloud.rcap", file="testData/testConfig.json")
rcapConfig <- jsonlite::fromJSON(jsonFile, simplifyVector = FALSE)
rcapFunctions <- parseConfig(rcapConfig)
}

