% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseControl}
\alias{parseControl}
\title{Parse a control for OCAPs}
\usage{
parseControl(rcapControl)
}
\arguments{
\item{rcapControl}{List containing the control object from the site JSON.}
}
\value{
The name of the R function that is callable from the client-side (OCAP)

}
\description{
Parse a control for OCAPs
}
\examples{
\dontrun{
pageFunctions <- lapply(rcapPage$controls, parseControl)
}

}

