% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parsePage}
\alias{parsePage}
\title{Parse an RCAP page for functions}
\usage{
parsePage(rcapPage)
}
\arguments{
\item{rcapPage}{a list containing the controls on a page}
}
\value{
list of the functions on the page and the functions of child pages in
a tree structure.

}
\description{
Process and return each control that creates an OCAP R function 
on a page. This function will call itself recursively for all child pages in
a depth first search.

}
\examples{
\dontrun{
pageFunctions <- lapply(rcapConfig$pages, parsePage)
}
}

