% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_f.R
\name{plot_hist}
\alias{plot_hist}
\title{plot_hist}
\usage{
plot_hist(x, logger = NULL)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- plot_hist(1)  
}
}
