% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_data.R
\name{plot_NA_cond}
\alias{plot_NA_cond}
\title{Title Create a histogram/bar plot of a variable conditional on missingness in another variable}
\usage{
plot_NA_cond(dat, x, z)
}
\arguments{
\item{dat}{An incomplete dataset of class dataframe}

\item{x}{A variable to plot}

\item{z}{A second variable to plot}
}
\value{
A ggplot object with two facets (variable z observed vs. missing)
}
\description{
Title Create a histogram/bar plot of a variable conditional on missingness in another variable
}
