% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_03_convergence.R
\name{plot_rhat}
\alias{plot_rhat}
\title{Title Plot non-convergence diagnostic 'R hat'/potential scale reduction factor/Gelman-Rubin statistic}
\usage{
plot_rhat(imp, x, theta = "means")
}
\arguments{
\item{imp}{A multiply imputed data set (mids) object}

\item{x}{A variable to plot}

\item{theta}{The parameter to plot (either "means" for chain means, or "vars" for chain variances)}
}
\value{
A ggplot object visualizing non-convergence
}
\description{
Title Plot non-convergence diagnostic 'R hat'/potential scale reduction factor/Gelman-Rubin statistic
}
