% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_03_convergence.R
\name{plot_trace}
\alias{plot_trace}
\title{Title Plot convergence parameters for a single variable}
\usage{
plot_trace(d, x)
}
\arguments{
\item{d}{A dataframe with preprocessed imputation chains generated via the function `preprocess_thetas()`}

\item{x}{A variable to plot}
}
\value{
A ggplot object visualizing a traceplot
}
\description{
Title Plot convergence parameters for a single variable
}
