% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_04_imputations.R
\name{plot_xy}
\alias{plot_xy}
\title{Title Plot scatterplot of imputed data}
\usage{
plot_xy(imp, x, y)
}
\arguments{
\item{imp}{A multiply imputed data set (`mids`) object}

\item{x}{A variable to plot}

\item{y}{A second variable to plot}
}
\value{
A ggplot object to replace xyplot
}
\description{
Title Plot scatterplot of imputed data
}
