% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processControl.R, R/processControl.interactiveplot.R, R/processControl.rplot.R
\name{processControl}
\alias{processControl}
\alias{processControl.interactiveplot}
\alias{processControl.rplot}
\title{Process a control for OCAPs}
\usage{
processControl(rcapControl)

\method{processControl}{interactiveplot}(rcapControl)

\method{processControl}{rplot}(rcapControl)
}
\arguments{
\item{rcapControl}{A list version of the control JSON object}
}
\value{
The function name of the OCAP or an empty list
}
\description{
A generic function. The S3 class for a control should derive from
its \code{type} property from the JSON file. This function will then call
different methods for each type of control. Controls that do not require OCAPs
will not have a method and the deafult behaviour is to do nothing and return
an empty list. Controls that create OCAPs will generate the R function and
source to the global envirnoment before returning the name of the R function.
}
\examples{
\dontrun{
# Apply the class from the type if available
if (!is.null(rcapControl$type)) {
  attr(rcapControl, "class") <- rcapControl$type
}

processControl(rcapControl)
}
}

