% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_object_functions.R
\name{query_feature_values_with_cohort_object}
\alias{query_feature_values_with_cohort_object}
\title{Title}
\usage{
query_feature_values_with_cohort_object(
  cohort_object,
  feature = NA,
  class = NA
)
}
\arguments{
\item{class}{}
}
\description{
Title
}
