% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcap.result.R
\name{rcap.result}
\alias{rcap.result}
\title{Load site configuration and setup OCAPs}
\usage{
rcap.result(rcapConfigFileName = "rcap_designer.json", inline = FALSE)
}
\arguments{
\item{rcapConfigFileName}{Character vector: The name of the json file in the assets.}

\item{inline}{logical: If TRUE this will attempt to call rcw.result with the rcw.inline wrapper}
}
\description{
Load site configuration and setup OCAPs
}
\examples{
\dontrun{
rcap.result("rcap_designer.json")
}
}

