% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcapSessionInfo}
\alias{rcapSessionInfo}
\title{Return relevant session info in a named list}
\usage{
rcapSessionInfo()
}
\value{
A list with username and nodename (host)
}
\description{
This information is returned to the front end
}

