% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocaps.R
\name{rcloud.rcap.global.functions}
\alias{rcloud.rcap.global.functions}
\title{Find all functions in global environment}
\usage{
rcloud.rcap.global.functions()
}
\value{
A character vector of function names
}
\description{
Find all functions in global environment
}
\examples{
f <- function(x) {x}
g <- "not a function"
rcloud.rcap.global.functions()
}

