% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcloudEnv}
\alias{rcloudEnv}
\title{The environment the notebook and mini.html are run in}
\usage{
rcloudEnv()
}
\value{
An environment.
}
\description{
This environment is used to search for functions that are used
to update the dashboard, and for variables that are associated
with dashboard controls.
}
\details{
This is currently the global environment.
}

