% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_01_data.R
\name{read_data}
\alias{read_data}
\title{Title Upload a file with incomplete data, or don't (then use mice::boys)}
\usage{
read_data(file)
}
\arguments{
\item{file}{An incomplete dataset with file extension '.csv', '.tsv', '.rda', or '.RData'}
}
\value{
The file loaded in the working directory as dataframe
}
\description{
Title Upload a file with incomplete data, or don't (then use mice::boys)
}
