% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{robyn_learn}
\alias{robyn_learn}
\title{Robyn's Educational UI (RobynLearn)}
\usage{
robyn_learn(...)
}
\arguments{
\item{...}{Additional parameters}
}
\description{
Using Shiny on top of Robyn users are enabled to both learn Robyn
in a more efficient way, as well as develop end-to-end preliminary
models with limited flexibility. All features of Robyn are represented
in some capacity in the UI, and it remains the best way for those with
limited Marketing Mix Modeling and/or coding experience to familiarize
themselves with Robyn & MMM.
}
\author{
Kyle Goldberg (kylegoldberg@fb.com)
}
