% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{package}
\name{RobynLearn}
\alias{RobynLearn}
\alias{RobynLearn-package}
\title{Robyn Learn}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Using Shiny on top of Robyn users are enabled to both learn Robyn in a more efficient way, as well as develop end-to-end preliminary models with limited flexibility. All features of Robyn are represented in some capacity in the UI, and it remains the best way for those with limited Marketing Mix Modeling and/or coding experience to familiarize themselves with Robyn & MMM.
}
\author{
\strong{Maintainer}: Kyle Goldberg \email{kylegoldberg@fb.com}

Authors:
\itemize{
  \item Rotem Blinder \email{rblinder@fb.com}
  \item Bernardo Lares \email{bernardolares@fb.com}
  \item Gufeng Zhou \email{gufeng@fb.com}
  \item Leonel Sentana \email{leonelsentana@fb.com}
  \item Igor Skokan \email{igorskokan@fb.com}
  \item Ruiyao Li \email{ruiyaoli@fb.com}
}

Other contributors:
\itemize{
  \item Meta Platforms, Inc. [copyright holder, funder]
}

}
