% Generated by roxygen2 (5.0.0.9000): do not edit by hand
% Please edit documentation in R/controlTypes.R
\docType{data}
\name{RPlotControl}
\alias{RPlotControl}
\title{Control class for an R plot}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fbef32fbd50> 
 - attr(*, "name")= chr "RPlotControl_generator"
}}
\usage{
RPlotControl
}
\description{
It runs an R function to update the plot, and then pushes
the new plot to the front-end.
}
\keyword{datasets}

