% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_async_f.R
\name{select_async}
\alias{select_async}
\title{select_async}
\usage{
select_async(x, async = FALSE)
}
\arguments{
\item{x}{x}

\item{async}{logical. Whether the function is being run asynchronously}
}
\description{
Does x
}
\examples{
{
y <- select_async(1)  
}
}
