% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_query_f.R
\name{select_query}
\alias{select_query}
\title{select_query}
\usage{
select_query(x, logger = NULL)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- select_query(1)  
}
}
