% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{sgt_assuntos}
\alias{sgt_assuntos}
\title{sgt_assuntos}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 2 columns.
}
\usage{
sgt_assuntos
}
\description{
sgt_assuntos
}
\keyword{datasets}
