% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{sgt_classes}
\alias{sgt_classes}
\title{sgt_classes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 788 rows and 5 columns.
}
\usage{
sgt_classes
}
\description{
sgt_classes
}
\keyword{datasets}
