% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{sgt_movs}
\alias{sgt_movs}
\title{sgt_movs}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 693 rows and 4 columns.
}
\usage{
sgt_movs
}
\description{
sgt_movs
}
\keyword{datasets}
