% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sourceControlCode.R
\name{sourceControlCode}
\alias{sourceControlCode}
\title{Evaluate OCAP functions into global namespace}
\usage{
sourceControlCode(funText, controlId)
}
\arguments{
\item{funText}{Character vector of each line of the function to be sourced}

\item{controlId}{The control ID, as it is in the json so that errors can be reported}
}
\value{
Nothing
}
\description{
Evaluate OCAP functions into global namespace
}
\examples{
\dontrun{
funText <- fillTemplate("rplotTemplate.R",
                        list(control.id=rcapControl$id,
                        control.controlProperties.1=rcapControl$controlProperties[[1]]$value))

sourceControlCode(funText, rcapControl$id)
}
}

