% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_feature_string}
\alias{transform_feature_string}
\title{Transform Feature String}
\usage{
transform_feature_string(feature, transformation)
}
\arguments{
\item{feature}{A string, the name of the feature}

\item{transformation}{A string, the name of the transformation}
}
\description{
Transform Feature String
}
