% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_filter_f.R
\name{transform_filter}
\alias{transform_filter}
\title{transform_filter}
\usage{
transform_filter(x, logger = NULL)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- transform_filter(1)  
}
}
