% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.r
\docType{class}
\name{zscore-class}
\alias{Zscore}
\alias{zscore-class}
\title{Class for calculationg z-score
constructor}
\arguments{
\item{all}{Metrics data of population}

\item{target}{Metrics data of target project for calculationg z-socre}

\item{params}{Parameters of metrics 
MIN/MAX value for rendering RadarChart}
}
\value{
data.frame having parameter name as header and each value is static c(-3, 3)
Calculate z-score displayed as RadarChart

z-score
}
\description{
Class for calculationg z-score
constructor
}

