% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_metrics.r
\docType{class}
\name{zscore_metrics-class}
\alias{ZscoreMetrics}
\alias{zscore_metrics-class}
\title{偏差値を測定するメトリクスに関するクラス
コンストラクタ}
\arguments{
\item{n}{このメトリクスの名前}

\item{f}{メトリクスを導出するための関数 (E.g.: x$code_smells / x$lines)}
}
\description{
偏差値を測定するメトリクスに関するクラス
コンストラクタ
}

