% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunomodulators_functions.R
\name{build_im_dt_tbl}
\alias{build_im_dt_tbl}
\title{Build Immunomodulators Datatable Tibble}
\usage{
build_im_dt_tbl(tbl)
}
\arguments{
\item{tbl}{A tibble}
}
\description{
Build Immunomodulators Datatable Tibble
}
