% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlTypes.R
\docType{data}
\name{control_classes}
\alias{control_classes}
\title{Front-end control types and matching back-end classes}
\format{An object of class \code{list} of length 18.}
\usage{
control_classes
}
\description{
Front-end control types and matching back-end classes
}
\keyword{datasets}

