% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getControls.R
\name{getControls}
\alias{getControls}
\title{Return a list of all controls}
\usage{
getControls(rcapConfig)
}
\arguments{
\item{rcapConfig}{a list generated by jsonlite::fromJSON(json, simplifyVector = TRUE)}
}
\value{
List of controls (each a list)
}
\description{
Return a list of all controls
}

