% Generated by roxygen2 (5.0.0.9000): do not edit by hand
% Please edit documentation in R/getControls.R
\name{getControls}
\alias{getControls}
\title{Return a list of all controls}
\usage{
getControls(rcapConfig)
}
\arguments{
\item{rcapConfig}{a list generated by
\code{jsonlite::fromJSON}, with \code{simplifyVector = TRUE}.}
}
\value{
List of controls (each a list itself).
}
\description{
The pages of the dashboard and the controls form a hierarchy,
as controls can have child controls. This parser goes over the
hierarchy recursively, and returns a flat list containing
all controls.
}
\details{
Each list element corresponds to a control, it is a list itself,
with the followinf members:
\itemize{
\item \code{type} see \code{control_classes} for the types.
\item \code{x}, \code{y} position on the grid.
\item \code{width}, \code{height} size on te grid.
\item \code{id} unique identifier, this is used in the protocol.
\item \code{controlProperties} another list, with \code{uid},
code{value} pairs, containing the configuration of the control.
(E.g. labels, associated R variables, or R function, etc.)
\item \code{childControls} if a control has children, they are
recursively included here.
}

See \preformatted{  system.file("tests", "testthat",
"testConfig.json", package = "rcloud.rcap")
} for an example JSON file describing controls.
}

