% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcap.result.R
\name{rcap.result}
\alias{rcap.result}
\title{Read rcap configuration and create OCAPs}
\usage{
rcap.result(rcapConfigFileName = "rcap_designer.json", inline = FALSE)
}
\arguments{
\item{rcapConfigFileName}{Character vector: The name of the json file in the assets.}

\item{inline}{logical: If TRUE this will attempt to call rcw.result with the rcw.inline wrapper}
}
\description{
Load site configuration from the json file in the notebook assests.
Parse the controls for any functions that need creating and and setup the OCAPs.
This function is a wrapper around \code{rcloud.web::rcw.result}.

}
\examples{
\dontrun{
rcap.result("rcap_designer.json")
}

}

