% Generated by roxygen2 (5.0.0.9000): do not edit by hand
% Please edit documentation in R/rcap.result.R
\name{rcap.result}
\alias{rcap.result}
\title{Set up a dashboard}
\usage{
rcap.result(rcapConfigFileName = "rcap_designer.json")
}
\arguments{
\item{rcapConfigFileName}{Character vector:
The name of the json file in the assets.}
}
\description{
This function must called from the last cell of the notebook.
When running in mini.html, it requests the asset that contains the
configuration of the dashboard, and initializes the viewer from it.
}
\details{
Then it creates the server side controller, and the controller
creates all the control objects that corresponds to the elements
of the dashboard.

This function should not be called from the notebook itself.
}

