% Generated by roxygen2 (5.0.0.9000): do not edit by hand
% Please edit documentation in R/ocaps.R
\name{rcloud.rcap.global.functions}
\alias{rcloud.rcap.global.functions}
\title{Find all functions in the rcloud environment}
\usage{
rcloud.rcap.global.functions()
}
\value{
A character vector of function names
}
\description{
The rcloud environment is the environment the notebook
runs in. This is currently the global environment, i.e. .GlobalEnv.
}
\details{
This function is used by the designer, to provide suggestions
whenever the designer user needs to supply an R function name;
E.g. when creating an R plot, the function that does the plotting,
or when a dropdown is populated by an R function.
}

