% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlTypes.R
\docType{data}
\name{RPlotControl}
\alias{RPlotControl}
\title{Control class for an R plot}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RPlotControl
}
\description{
It runs an R function to update the plot, and then pushes
the new plot to the front-end.
}
\keyword{datasets}

